+++ src/model/interface/node.ts
@@ -1,5 +1,6 @@
 import { Console } from "@/common/Console";
 import { DatabaseType, ModelType } from "@/common/constants";
+import { getKey } from "@/common/state";
 import { Util } from "@/common/util";
 import { DbTreeDataProvider } from "@/provider/treeDataProvider";
 import { getSqliteBinariesPath } from "@/service/connect/sqlite/sqliteCommandValidation";
@@ -7,6 +8,7 @@ import { ConnectionManager } from "@/service/connectionManager";
 import { SqlDialect } from "@/service/dialect/sqlDialect";
 import { QueryUnit } from "@/service/queryUnit";
 import { ServiceManager } from "@/service/serviceManager";
+import { platform } from "os";
 import * as vscode from "vscode";
 import { Memento } from "vscode";
 var commandExistsSync = require('command-exists').sync;
@@ -36,7 +38,7 @@ export abstract class Node extends vscode.TreeItem implements CopyAble {
     public connectTimeout?: number;
     public requestTimeout?: number;
     public includeDatabases?: string;
-
+    public connectionUrl?: string;
     /**
      * ssh
      */
@@ -52,7 +54,7 @@ export abstract class Node extends vscode.TreeItem implements CopyAble {
     public disable?: boolean;
 
     /**
-     * context
+     * using to distingush connectHolder, childCache, elementState
      */
     public uid: string;
     public key: string;
@@ -61,6 +63,7 @@ export abstract class Node extends vscode.TreeItem implements CopyAble {
     public parent?: Node;
 
     public useSSL?: boolean;
+    public caPath?: string;
     public clientCertPath?: string;
     public clientKeyPath?: string;
 
@@ -81,6 +84,12 @@ export abstract class Node extends vscode.TreeItem implements CopyAble {
      * es only
      */
     public scheme: string;
+    public esAuth: string;
+    public esToken: string;
+    /**
+     * using when ssh tunnel
+     */
+    public esUrl: string;
 
     /**
      * encoding, ftp only
@@ -100,7 +109,6 @@ export abstract class Node extends vscode.TreeItem implements CopyAble {
         this.port = source.port
         this.user = source.user
         this.password = source.password
-        if (!this.database) this.database = source.database
         this.timezone = source.timezone
         this.useSSL = source.useSSL
         this.clientCertPath = source.clientCertPath
@@ -108,14 +116,14 @@ export abstract class Node extends vscode.TreeItem implements CopyAble {
         this.ssh = source.ssh
         this.usingSSH = source.usingSSH
         this.scheme = source.scheme
+        this.esAuth = source.esAuth
+        this.esToken = source.esToken
         this.encoding = source.encoding
         this.showHidden = source.showHidden
-        if (!this.schema) {
-            this.schema = source.schema
-        }
         this.connectionKey = source.connectionKey
         this.global = source.global
         this.dbType = source.dbType
+        this.connectionUrl = source.connectionUrl
         if (source.connectTimeout) {
             this.connectTimeout = parseInt(source.connectTimeout as any)
             source.connectTimeout = parseInt(source.connectTimeout as any)
@@ -131,6 +139,8 @@ export abstract class Node extends vscode.TreeItem implements CopyAble {
         this.authType = source.authType
         this.disable = source.disable
         this.includeDatabases = source.includeDatabases
+        if (!this.database) this.database = source.database
+        if (!this.schema) this.schema = source.schema
         if (!this.provider) this.provider = source.provider
         if (!this.context) this.context = source.context
         // init dialect
@@ -140,7 +150,7 @@ export abstract class Node extends vscode.TreeItem implements CopyAble {
         if (this.disable) {
             this.command = { command: "mysql.connection.open", title: "Open Connection", arguments: [this] }
         }
-        this.key = source.key;
+        this.key = source.key || this.key;
         this.initUid();
         // init tree state
         this.collapsibleState = DatabaseCache.getElementState(this)
@@ -159,9 +169,9 @@ export abstract class Node extends vscode.TreeItem implements CopyAble {
     public async indent(command: IndentCommand) {
 
         try {
-            const connectionKey = command.connectionKey || this.connectionKey;
+            const connectionKey = getKey(command.connectionKey || this.connectionKey);
             const connections = this.context.get<{ [key: string]: Node }>(connectionKey, {});
-            const key = this.key 
+            const key = this.key
 
             switch (command.command) {
                 case CommandKey.add:
@@ -244,7 +254,7 @@ export abstract class Node extends vscode.TreeItem implements CopyAble {
     public getConnectId(opt?: SwitchOpt): string {
 
 
-        let uid = (this.usingSSH) ? `${this.ssh.host}@${this.ssh.port}` : `${this.host}@${this.instanceName?this.instanceName:this.port}`;
+        let uid = (this.usingSSH) ? `${this.ssh.host}@${this.ssh.port}` : `${this.host}@${this.instanceName ? this.instanceName : this.port}`;
 
         uid = `${this.key}@@${uid}`
 
@@ -286,16 +296,19 @@ export abstract class Node extends vscode.TreeItem implements CopyAble {
             command = `mysql -u ${this.user} -p${this.password} -h ${this.host} -P ${this.port} \n`;
         } else if (this.dbType == DatabaseType.PG) {
             this.checkCommand('psql');
-            command = `set "PGPASSWORD=${this.password}" && psql -U ${this.user} -h ${this.host} -p ${this.port} \n`;
-        }else if(this.dbType==DatabaseType.REDIS){
+            let prefix = platform() == 'win32' ? 'set' : 'export';
+            command = `${prefix} "PGPASSWORD=${this.password}" && psql -U ${this.user} -h ${this.host} -p ${this.port} -d ${this.database} \n`;
+        } else if (this.dbType == DatabaseType.REDIS) {
             this.checkCommand('redis-cli');
-            command = `redis-cli -h ${this.host} -p ${this.port} \n`;   
-        }else if(this.dbType==DatabaseType.MONGO_DB){
+            command = `redis-cli -h ${this.host} -p ${this.port} \n`;
+        } else if (this.dbType == DatabaseType.MONGO_DB) {
             this.checkCommand('mongo');
-            command = `mongo --host ${this.host} --port ${this.port} ${this.user&&this.password?` -u ${this.user} -p ${this.password}`:''} \n`;   
-        }else if(this.dbType==DatabaseType.SQLITE){
-            
-            command = `${getSqliteBinariesPath()} ${this.dbPath} \n`;   
+            command = `mongo --host ${this.host} --port ${this.port} ${this.user && this.password ? ` -u ${this.user} -p ${this.password}` : ''} \n`;
+        } else if (this.dbType == DatabaseType.SQLITE) {
+            command = `${getSqliteBinariesPath()} ${this.dbPath} \n`;
+        } else {
+            vscode.window.showErrorMessage(`Database type ${this.dbType} not support open terminal.`)
+            return;
         }
         const terminal = vscode.window.createTerminal(this.dbType.toString())
         terminal.sendText(command)