+++ src/provider/tableInfoHoverProvider.ts
@@ -12,9 +12,24 @@ export class TableInfoHoverProvider implements HoverProvider {
 
         const sourceCode = await tableNode?.execute<any[]>(tableNode.dialect.showTableSource(tableNode.schema, tableNode.table))
         if (sourceCode) {
-            const markdownStr = new vscode.MarkdownString();
+            const args = [{ sql: `SELECT * FROM ${tableNode.table}` }];
+            const runCommandUri = vscode.Uri.parse(`command:mysql.runQuery?${encodeURIComponent(JSON.stringify(args))}`);
+            const markdownStr = new vscode.MarkdownString(`[Query Table](${runCommandUri})`);
+            markdownStr.isTrusted=true;
             markdownStr.appendCodeblock(sourceCode[0]['Create Table'], "sql");
             return new vscode.Hover(markdownStr);
+        }
+
+        const selections = vscode.window.activeTextEditor?.selections || []
+        for (const selection of selections) {
+            if (selection.contains(position)) {
+                const args = [{ sql: document.getText(selection) }];
+                const runCommandUri = vscode.Uri.parse(`command:mysql.runQuery?${encodeURIComponent(JSON.stringify(args))}`);
+                const contents = new vscode.MarkdownString(`[Run Selected SQL](${runCommandUri})`);
+                contents.isTrusted = true;
+                const hover = new vscode.Hover(contents);
+                return hover;
+            }
         }
 
         return null;